move args to a separate file
